added about page css -- forgot to add to last commit refactored pages
