+++ app/pages/Stats.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router';
 
 import axios from 'axios';
 
-import Table from './Stats/Table';
+import Table from '../components/Stats/Table';
 import data from '../data/github';
 
 class Stats extends Component {
