+++ public/css/layout/_about.scss
@@ -1,0 +1,8 @@
+// About page (/about)
+
+#about {
+
+  p {
+    margin: auto;
+  }
+}
